+++ starkbank/merchantpurchase/merchant_purchase.go
@@ -21,6 +21,7 @@ type MerchantPurchase struct {
 	HolderEmail        string                 `json:",omitempty"`
 	HolderPhone        string                 `json:",omitempty"`
 	HolderId           string                 `json:",omitempty"`
+	SoftDescriptor     string                 `json:",omitempty"`	
 	FundingType        string                 `json:",omitempty"`
 	BillingCountryCode string                 `json:",omitempty"`
 	BillingCity        string                 `json:",omitempty"`