+++ starkbank/merchantsession/purchase.go
@@ -16,6 +16,7 @@ type Purchase struct {
 	HolderEmail        string                 `json:",omitempty"`
 	HolderPhone        string                 `json:",omitempty"`
 	HolderId		   string                 `json:",omitempty"`
+	SoftDescriptor     string                 `json:",omitempty"`
 	FundingType        string                 `json:",omitempty"`
 	BillingCountryCode string                 `json:",omitempty"`
 	BillingCity        string                 `json:",omitempty"`