+++ CHANGELOG.md
@@ -14,8 +14,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 ### Added
-- HolderId attribute to MerchantSession and MerchantPurchase
-- SoftDescriptor attribute to MerchantSession
+- HolderId and SoftDescriptor attributes to MerchantSession and MerchantPurchase
 
 ## [1.5.0] - 2026-01-13
 ### Added